Version 1.0.0
Game logic has been created allowing the user to actually play the game.
Changed the value variable from a String to an int to reduce the
conversions between the two. Removed the "Check Solution" button since
the puzzle in not stored in a complete state anywhere. Can be added back
in but should probably be turned into hints. Added getter/setters to
allow for better encapsulation of the tiles.